back-end/requirements.txt
Flask
python-dotenv
Flask-CORS
langchain-core
langchain-openai
langchain-anthropic
langchain-google-genai
langchain-community
